extract load function
