+++ Example/Snabble/AppDelegate.swift
@@ -133,24 +133,3 @@ extension AppDelegate: CheckInManagerDelegate {
         shopsViewController?.viewModel.shop = nil
     }
 }
-
-func loadJSON<T: Decodable>(_ filename: String) -> T {
-    let data: Data
-
-    guard let file = Bundle.main.url(forResource: filename, withExtension: "json") else {
-        fatalError("Couldn't find \(filename) in bundle.")
-    }
-
-    do {
-        data = try Data(contentsOf: file)
-    } catch {
-        fatalError("Couldn't load \(filename) from bundle:\n\(error)")
-    }
-
-    do {
-        let decoder = JSONDecoder()
-        return try decoder.decode(T.self, from: data)
-    } catch {
-        fatalError("Couldn't parse \(filename) as \(T.self):\n\(error)")
-    }
-}
