+++ Example/Snabble/LoadJSON.swift
@@ -1,0 +1,30 @@
+//
+//  LoadJSON.swift
+//  SnabbleSampleApp
+//
+//  Created by Andreas Osberghaus on 25.08.22.
+//  Copyright © 2022 snabble. All rights reserved.
+//
+
+import Foundation
+
+func loadJSON<T: Decodable>(_ filename: String) -> T {
+    let data: Data
+
+    guard let file = Bundle.main.url(forResource: filename, withExtension: "json") else {
+        fatalError("Couldn't find \(filename) in bundle.")
+    }
+
+    do {
+        data = try Data(contentsOf: file)
+    } catch {
+        fatalError("Couldn't load \(filename) from bundle:\n\(error)")
+    }
+
+    do {
+        let decoder = JSONDecoder()
+        return try decoder.decode(T.self, from: data)
+    } catch {
+        fatalError("Couldn't parse \(filename) as \(T.self):\n\(error)")
+    }
+}
